+++ src/main/java/rappsilber/ms/sequence/digest/Digestion.java
@@ -713,20 +713,13 @@ public class Digestion {
             return (Digestion) m.invoke(null, Options);
 
 
-        } catch (ClassNotFoundException ex) {
+        } catch (ClassNotFoundException|IllegalAccessException|IllegalArgumentException|
+                InvocationTargetException|NoSuchMethodException|SecurityException ex) {
             Logger.getLogger(Digestion.class.getName()).log(Level.SEVERE, null, ex);
-        } catch (IllegalAccessException ex) {
-            Logger.getLogger(Digestion.class.getName()).log(Level.SEVERE, null, ex);
-        } catch (IllegalArgumentException ex) {
-            Logger.getLogger(Digestion.class.getName()).log(Level.SEVERE, null, ex);
-        } catch (InvocationTargetException ex) {
-            Logger.getLogger(Digestion.class.getName()).log(Level.SEVERE, null, ex);
-        } catch (NoSuchMethodException ex) {
-            Logger.getLogger(Digestion.class.getName()).log(Level.SEVERE, null, ex);
-        } catch (SecurityException ex) {
-            Logger.getLogger(Digestion.class.getName()).log(Level.SEVERE, null, ex);
+            if (ex.getCause() != null)
+                Logger.getLogger(Digestion.class.getName()).log(Level.SEVERE, "caused by " , ex.getCause());
+            throw new Error(ex);
         }
-        return null;
 
     }
 
@@ -739,18 +732,12 @@ public class Digestion {
             return (Digestion) m.invoke(null, Options, config);
 
 
-        } catch (ClassNotFoundException ex) {
+        } catch (ClassNotFoundException|IllegalAccessException|IllegalArgumentException|
+                InvocationTargetException|NoSuchMethodException|SecurityException ex) {
             Logger.getLogger(Digestion.class.getName()).log(Level.SEVERE, null, ex);
-        } catch (IllegalAccessException ex) {
-            Logger.getLogger(Digestion.class.getName()).log(Level.SEVERE, null, ex);
-        } catch (IllegalArgumentException ex) {
-            Logger.getLogger(Digestion.class.getName()).log(Level.SEVERE, null, ex);
-        } catch (InvocationTargetException ex) {
-            Logger.getLogger(Digestion.class.getName()).log(Level.SEVERE, null, ex);
-        } catch (NoSuchMethodException ex) {
-            Logger.getLogger(Digestion.class.getName()).log(Level.SEVERE, null, ex);
-        } catch (SecurityException ex) {
-            Logger.getLogger(Digestion.class.getName()).log(Level.SEVERE, null, ex);
+            if (ex.getCause() != null)
+                Logger.getLogger(Digestion.class.getName()).log(Level.SEVERE, "caused by " , ex.getCause());
+            System.exit(-1);
         }
         return null;