+++ Changelog.md
@@ -1,3 +1,12 @@
+
+Version 1.8.8
+  * starting a search without defined in and output files gives a better response
+  * BugFix: correct the behaviour if a modification is defined nterm as part of the specificity
+  * BugFix: mzML reader not assigning titel to spectra
+
+Version 1.8.7
+  * skyline converter writes crosslinked ssl file
+
 Version 1.8.6
   * xiSEARCH version gets forwarded to xiFDR (if >= 2.3.1)
   * BugFix: linear peptide matches where exported with two columns missing (from the empty peptide2 columns)